backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
redis==5.0.8
minio==7.2.8
httpx==0.27.2
sib-api-v3-sdk==7.6.0
python-dotenv==1.0.1
ruff==0.6.0
Pillow==10.4.0
